+++ tests/run-tests.js
@@ -1,5 +1,6 @@
 import './sessionDefaults.test.js';
 import './scoringEngine.test.js';
+import './cardRenderer.test.js';
 import './sessionSetup.test.js';
 import './sessionActions.test.js';
 import './fullCatalogIntegration.test.js';